Update index.html
If you select ^ close on the debut year option, then obviously you only show 5 years above. However I want it to be that if you select ^ above, then it shows all years ^ but not within 5 years (because if the answer was within 5 years then it would’ve said close). Same thing for v obviously. As well as for the popularity you just make it 50.